requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-docx==1.1.0
Pillow==10.1.0
bcrypt==4.1.2
python-multipart==0.0.6
aiofiles==23.2.1 